Move art opcodes out of animation.c/h, update SFALL_COMPATIBILITY
